corrections fonctions laser & asserv_angle

+++ navigation/approche_finale/src/Point.cpp
@@ -1,6 +1,6 @@
 #include "Point.h"
 
-Point::Point(float r, float phi){
+Point::Point(float r, double phi){
 m_r = r;
 m_phi = phi;
 }
